EE-9765 Code review change: moved Individual into new application.domain package
