+++ src/main/java/uk/gov/digital/ho/pttg/application/IncomeSummaryContext.java
@@ -4,6 +4,7 @@ import lombok.Getter;
 import lombok.experimental.Accessors;
 import org.springframework.hateoas.Link;
 import org.springframework.hateoas.Resource;
+import uk.gov.digital.ho.pttg.application.domain.Individual;
 import uk.gov.digital.ho.pttg.dto.*;
 
 import java.util.List;
