+++ src/main/java/uk/gov/digital/ho/pttg/dto/EmbeddedIndividual.java
@@ -2,6 +2,7 @@ package uk.gov.digital.ho.pttg.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import uk.gov.digital.ho.pttg.application.domain.Individual;
 
 @AllArgsConstructor
 @Getter
