+++ src/test/java/uk/gov/digital/ho/pttg/application/HmrcClientTest.java
@@ -7,6 +7,7 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 import org.springframework.hateoas.Link;
 import org.springframework.test.util.ReflectionTestUtils;
+import uk.gov.digital.ho.pttg.application.domain.Individual;
 import uk.gov.digital.ho.pttg.dto.*;
 
 import java.math.BigDecimal;
