+++ src/test/java/uk/gov/digital/ho/pttg/application/domain/IndividualTest.java
@@ -1,7 +1,8 @@
-package uk.gov.digital.ho.pttg.dto;
+package uk.gov.digital.ho.pttg.application.domain;
 
 import org.junit.Test;
 import uk.gov.digital.ho.pttg.application.ApplicationExceptions.InvalidIdentityException;
+import uk.gov.digital.ho.pttg.application.domain.Individual;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
